requirements.txt
aiohttp==3.7.4.post0
alembic==1.8.1
annotated-types==0.7.0
anyio==3.6.1
apache-airflow==2.4.1
apache-airflow-providers-amazon==8.0.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-sqlite==3.2.1
apache-airflow-providers-vertica==3.3.1
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
boto3==1.26.137
botocore==1.29.137
bson==0.5.10
cachelib==0.9.0
cattrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.5
colorlog==4.8.0
comm==0.1.3
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==36.0.2
dags==0.2.3
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
distro-info==1.0
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.0rc9
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.2
Flask-AppBuilder==4.1.3
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fqdn==1.5.1
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==4.12.0
inflection==0.5.1
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonpointer==2.3
jsonschema==4.16.0
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kafka-python==2.0.2
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==7.5.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==2.0.5
multidict==6.0.4
mypy-boto3-appflow==1.26.125
mypy-boto3-rds==1.26.136
mypy-boto3-redshift-data==1.26.109
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
ply==3.11
postgres==4.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.2
psycopg==3.2.9
psycopg2-binary==2.9.6
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pycparser==2.21
pycurl==7.43.0.6
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.13.0
PyGObject==3.38.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.4.0
python-apt==2.2.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
redshift-connector==2.0.910
requests==2.28.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.5.1
s3transfer==0.6.1
scramp==1.4.4
Send2Trash==1.8.2
serverhub-agent==0.0.18
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.27
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
stack-data==0.6.2
supervisor==4.2.5
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.1.0
termcolor==2.0.1
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing-inspect==0.8.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
uc-micro-py==1.0.1
unattended-upgrades==0.1
unicodecsv==0.14.1
uri-template==1.2.0
urllib3==1.26.12
vertica-python==1.3.2
watchtower==2.0.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==2.3.3
yarl==1.9.2
zipp==3.8.1